+++ projects/ui/src/lib/components/typography/index.ts
@@ -1,0 +1,3 @@
+export * from './kui-text-tone.type';
+export * from './kui-text-variant.type';
+export * from './kui-text.directive';